Added the Login section

+++ frontend/src/core/LoginBox.js
@@ -1,16 +1,37 @@
 import React, { useEffect, useState } from "react";
 import TextField from "@material-ui/core/TextField";
 import Button from "@material-ui/core/Button";
-import { Redirect } from "react-router";
 import history from "../HistoryContainer/History";
+import axios from "axios";
 
 export default function LoginTable() {
   const [employeeId, setEmployeeId] = useState(null);
+  const [employeePassword, setEmployeePassword] = useState(null);
 
   const handleClick = () => {
     console.log(employeeId);
-    sessionStorage.setItem("employeeId" , employeeId);
-    history.push("/employee");
+    var apiUrl = "http://localhost:5000/employeeBasic/verifyEmpNo";
+    console.log(employeePassword);
+    axios.post(apiUrl,{ 
+        id: employeeId,
+        password: employeePassword
+      })
+    .then((repos) => {
+        console.log(repos);
+      if(repos.data['exist'] && repos.data['password']){
+        sessionStorage.setItem("employeeId" , employeeId);
+        sessionStorage.setItem("employeeDesignation" , repos.data['designation']);
+        history.push("/employee");
+        window.location.reload();
+      }
+      else if(repos.data['exist'] && repos.data['password'] == false){
+        alert("Wrong Password for Employee Id " + employeeId);
+      }
+    else{
+        alert("No employee with id " + employeeId);
+      }
+    });
+    
   }
   return (
     <div>
@@ -21,8 +42,19 @@ export default function LoginTable() {
         onChange={(event) => {
           setEmployeeId(event.target.value);
         }}
-        label="Enter employee Id"
+        label="Enter Your employee Id"
         type="text"
+        fullWidth
+      />
+      <TextField
+        autoFocus
+        margin="dense"
+        id="name"
+        onChange={(event) => {
+          setEmployeePassword(event.target.value);
+        }}
+        label="Enter Your PassWord"
+        type="password"
         fullWidth
       />
       <Button
